Add optimisation for fully populated chunks
If a chunk is filled using the 'fill' operation then mark it as full
(compressed) rather than uselessly populating all the voxels.
If a modification of a full+compressed chunk is required later it is
unpacked and the modification applied.
This change saves time and memory at the cost of a slight delay when
attempting to modify a full chunk for the first time.
Currently it is assumed voxels take only the values 0 (empty) or 1
(solid), this method will need to change later when new values are
introduced.