Implement collision checking
Collision detection and response in all axis.
Added AABB bounding box class with utility functions for collision
tests.
In AABB class there is an issue with floating point comparisons which
was resolved through the use of a small constant tolerance (epsilon),
this will need revisiting.
Gravity and collision detection can be enabled or disabled.